LPD-51750 Move utils directory to liferay bundle
So that everything is there and then we can extract main submodule's
exclusive code
