+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/utils/client_extensions/loadModule.ts
