+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/utils/date_time/constants.ts
